GPU Stat: cleanups
